Params and GUI both support sample rate multiplier

+++ include/vapor/RayCasterParams.h
@@ -47,8 +47,8 @@ public:
     //
     long   GetCastingMode() const;
     void   SetCastingMode( long );
-    double GetSampleMultiplier() const;
-    void   SetSampleRateMultiplier( double );
+    long   GetSampleMultiplier() const;     // ComboBox index is held here. Need to translate
+    void   SetSampleRateMultiplier( long ); //   to real multipliers in RayCaster.cpp
 
 protected:
